Split up fields & utils into subpackages to avoid dependency issues. Move Enum into enum module.
