+++ flask_common/fields/__init__.py
@@ -1,0 +1,32 @@
+# Basic fields that don't have any non-MongoEngine dependencies
+from .basic import LowerEmailField, LowerStringField, TrimmedStringField
+
+# Crypto fields
+try:
+    from .crypto import EncryptedStringField
+except ImportError:
+    pass
+
+# Phone numbers fields
+try:
+    from .phone import PhoneField
+except ImportError:
+    pass
+
+# Sorted set fields
+try:
+    from .sorted_set import ISortedSetField, SortedSetField
+except ImportError:
+    pass
+
+# Timezone fields
+try:
+    from .tz import TimezoneField
+except ImportError:
+    pass
+
+# UUID fields
+try:
+    from .id import IDField
+except ImportError:
+    pass
