+++ flask_common/fields/tz.py
@@ -1,0 +1,18 @@
+from mongoengine.fields import StringField
+import pytz
+
+
+class TimezoneField(StringField):
+    def __init__(self, *args, **kwargs):
+        defaults = {
+            'default': 'UTC',
+            'choices': tuple(zip(pytz.all_timezones, pytz.all_timezones)),
+        }
+        defaults.update(kwargs)
+        return super(TimezoneField, self).__init__(*args, **defaults)
+
+    def to_python(self, value):
+        return pytz.timezone(value)
+
+    def to_mongo(self, value):
+        return unicode(value)
