+++ flask_common/utils/__init__.py
@@ -1,0 +1,11 @@
+# UUID related helpers
+try:
+    from .id import *
+except ImportError:
+    pass
+
+# TODO: split these up
+try:
+    from .legacy import *
+except ImportError:
+    pass
