+++ flask_common/utils/legacy.py
@@ -28,6 +28,9 @@ from smtplib import SMTPDataError
 from socket import gethostname
 
 
+from ..enum import Enum # deprecated
+
+
 def returns_xml(f):
     @wraps(f)
     def decorated_function(*args, **kwargs):
@@ -191,39 +194,6 @@ def finite_float(value):
     if math.isnan(value) or math.isinf(value):
         raise ValueError("Can't convert %s to a finite float" % value)
     return value
-
-
-class Enum(object):
-    """
-    A list of constants that can be defined in a declarative way.
-
-    Example usage:
-
-    class MyEnum(Enum):
-        Choice1 = 'value1'
-        Choice2 = 'value2'
-
-    In this case, we can refer to the choices as MyEnum.Choice1 or
-    MyEnum.Choice2, and don't have to reference the actual string value, which
-    is prone to typos.
-    """
-
-    @classmethod
-    def values(cls):
-        """
-        Returns a list of all the values, e.g.: ('choice1', 'choice2')
-        """
-        return [getattr(cls,v) for v in dir(cls)
-                if not callable(getattr(cls,v)) and not v.startswith('_')]
-
-    @classmethod
-    def choices(cls):
-        """
-        Returns a list of choice tuples, e.g.:
-        [('value1', 'Choice1'), ('value2', 'Choice2')]
-        """
-        return [(getattr(cls,v), v) for v in dir(cls)
-                if not callable(getattr(cls,v)) and not v.startswith('_')]
 
 
 def grouper(n, iterable):
